refactor: move tag and author components to common folder
